mensage

+++ src/main/java/sv/com/cooperativa/models/dao/IAreaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import sv.com.cooperativa.models.entities.Area;
 
-public interface IAreaDao extends CrudRepository<Area, Long>{
+public interface IAreaDao extends CrudRepository<Area, Integer>{
 
 }
