+++ src/main/java/sv/com/cooperativa/models/dao/ICargoDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import sv.com.cooperativa.models.entities.Cargo;
 
-public interface ICargoDao extends CrudRepository<Cargo, Long> {
+public interface ICargoDao extends CrudRepository<Cargo, Integer> {
 
 }
