+++ src/main/java/sv/com/cooperativa/models/dao/ITasaInteresDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import sv.com.cooperativa.models.entities.TasaInteres;
 
-public interface ITasaInteresDao extends CrudRepository<TasaInteres, Long>{
+public interface ITasaInteresDao extends CrudRepository<TasaInteres, Integer>{
 
 }
