+++ src/main/java/sv/com/cooperativa/models/services/IArea.java
@@ -7,6 +7,6 @@ import sv.com.cooperativa.models.entities.Area;
 public interface IArea {
 	public List<Area> findAll();
 	public void Save(Area banco);
-	public Area findOne(Long id_area);
-	public void delete(Long id_area);
+	public Area findOne(Integer id_area);
+	public void delete(Integer id_area);
 }
