+++ src/main/java/sv/com/cooperativa/models/services/ICargo.java
@@ -7,6 +7,6 @@ import sv.com.cooperativa.models.entities.Cargo;
 public interface ICargo {
 	public List<Cargo> findAll();
 	public void Save(Cargo cargo);
-	public Cargo findOne(Long id_cargo);
-	public void delete(Long id_cargo);
+	public Cargo findOne(Integer id_cargo);
+	public void delete(Integer id_cargo);
 }
