+++ src/main/java/sv/com/cooperativa/models/services/ITasaInteres.java
@@ -7,6 +7,6 @@ import sv.com.cooperativa.models.entities.TasaInteres;
 public interface ITasaInteres {
 	public List<TasaInteres> findAll();
 	public void Save(TasaInteres tasa_interes);
-	public TasaInteres findOne(Long id_tasa_interes);
-	public void delete(Long id_tasa_interes);
+	public TasaInteres findOne(Integer id_tasa_interes);
+	public void delete(Integer id_tasa_interes);
 }
